reduce components
