--- apps/io-sign-backoffice-app/src/app/[locale]/[institution]/api-keys/create/_components/ApiKeyTestersSection.tsx
@@ -1,37 +1,0 @@
-"use client";
-
-import { useTranslations } from "next-intl";
-import { useFormContext, Controller } from "react-hook-form";
-
-import { Stack, Typography } from "@mui/material";
-
-import FiscalCodeListInput from "@/components/FiscalCodeListInput";
-
-import { FormFields } from "./CreateApiKeyForm";
-
-export default function ApiKeyNetworkSection() {
-  const t = useTranslations("firmaconio");
-  const { setValue } = useFormContext<FormFields>();
-  const onEditableListChange = (items: string[]) => {
-    setValue("testers", items);
-  };
-  return (
-    <Stack p={3} spacing={3} bgcolor="background.paper">
-      <Typography variant="h6">
-        {t("createApiKey.form.testers.title")}
-      </Typography>
-      <Typography variant="body1">
-        {t("createApiKey.form.testers.description")}
-      </Typography>
-      <Controller
-        name="testers"
-        render={({ field }) => (
-          <FiscalCodeListInput
-            items={field.value}
-            onChange={onEditableListChange}
-          />
-        )}
-      />
-    </Stack>
-  );
-}
