+++ apps/io-sign-backoffice-app/src/app/[locale]/[institution]/api-keys/create/_components/CreateApiKeyForm/index.tsx
@@ -1,0 +1,36 @@
+import { NextIntlClientProvider, useLocale, useMessages } from "next-intl";
+
+import { Institution } from "@/lib/selfcare/api";
+import { pick } from "lodash";
+
+import CreateApiKeyClientForm from "./CreateApiKeyClientForm";
+
+export default function CreateApiKeyForm({
+  children,
+  institution,
+}: {
+  children: React.ReactNode;
+  institution: Institution;
+}) {
+  const locale = useLocale();
+  const messages = useMessages();
+
+  if (!messages) {
+    throw new Error("unable to fetch localized messages");
+  }
+
+  const createApiKeyFormMessages = pick(messages, [
+    "firmaconio.apiKey.network",
+    "firmaconio.apiKey.testers",
+    "firmaconio.createApiKey.form",
+    "firmaconio.modals",
+  ]);
+
+  return (
+    <NextIntlClientProvider locale={locale} messages={createApiKeyFormMessages}>
+      <CreateApiKeyClientForm institution={institution}>
+        {children}
+      </CreateApiKeyClientForm>
+    </NextIntlClientProvider>
+  );
+}
