+++ apps/io-sign-backoffice-app/src/app/[locale]/[institution]/api-keys/create/page.tsx
@@ -1,28 +1,15 @@
 import PageHeader, { Props as PageHeaderProps } from "@/components/PageHeader";
-import { getConfig } from "@/config";
 import { SvgIconComponent, VpnKey } from "@mui/icons-material";
-import { Alert, Stack, Button } from "@mui/material";
+import { Alert, Stack } from "@mui/material";
 
-import { use } from "react";
+import { use, Suspense } from "react";
 
-import {
-  useTranslations,
-  NextIntlClientProvider,
-  useLocale,
-  useMessages,
-} from "next-intl";
-
-import { pick } from "lodash";
+import { useTranslations } from "next-intl";
 
 import CreateApiKeyForm from "./_components/CreateApiKeyForm";
-import ApiKeyDetailSection from "./_components/ApiKeyDetailSection";
-import ApiKeyNetworkSection from "./_components/ApiKeyNetworkSection";
-import ApiKeyTestersSection from "./_components/ApiKeyTestersSection";
-import FormActions from "./_components/FormActions";
+import ApiKeyMetadataCard from "./_components/ApiKeyMetadataCard";
 
-import NextLink from "next/link";
-import { Suspense } from "react";
-import { Institution, getInstitution } from "@/lib/selfcare/api";
+import { getInstitution } from "@/lib/selfcare/api";
 import ApiKeyEditableFieldCard from "./_components/ApiKeyEditableFieldCard";
 
 function CreateApiKeyPageHeader({
@@ -42,41 +29,12 @@ function CreateApiKeyPageHeader({
   return <PageHeader {...props} />;
 }
 
-function LocalizedCreateApiKeyForm({
-  children,
-  institution,
-}: {
-  children: React.ReactNode;
-  institution: Institution;
-}) {
-  const locale = useLocale();
-  const messages = useMessages();
-
-  if (!messages) {
-    throw new Error("unable to fetch localized messages");
-  }
-
-  const createApiKeyFormMessages = pick(messages, [
-    "firmaconio.apiKey.network",
-    "firmaconio.apiKey.testers",
-    "firmaconio.createApiKey.form",
-    "firmaconio.modals",
-  ]);
-
-  return (
-    <NextIntlClientProvider locale={locale} messages={createApiKeyFormMessages}>
-      <CreateApiKeyForm institution={institution}>{children}</CreateApiKeyForm>
-    </NextIntlClientProvider>
-  );
-}
-
 export default function CreateApiKeyPage({
   params,
 }: {
   params: { institution: string };
 }) {
   const t = useTranslations("firmaconio");
-  const { documentationUrl } = getConfig();
 
   const parent = {
     icon: VpnKey,
@@ -88,29 +46,24 @@ export default function CreateApiKeyPage({
 
   return (
     <Stack spacing={5} sx={{ maxWidth: "calc(100% - 300px)" }}>
-      <Suspense fallback={<span>loading...</span>}>
+      <Suspense>
         <CreateApiKeyPageHeader parent={parent} />
         <Alert
           severity="warning"
           variant="outlined"
           children={t.rich("createApiKey.alert")}
         />
-        <LocalizedCreateApiKeyForm institution={institution}>
-          <Stack spacing={5}>
-            <ApiKeyDetailSection documentationUrl={documentationUrl} />
-
-            <ApiKeyEditableFieldCard
-              field="cidrs"
-              i18n={{ namespace: "createApiKey.form.network" }}
-            />
-
-            <ApiKeyEditableFieldCard
-              field="testers"
-              i18n={{ namespace: "createApiKey.form.network" }}
-            />
-            <FormActions parent={parent} />
-          </Stack>
-        </LocalizedCreateApiKeyForm>
+        <CreateApiKeyForm institution={institution}>
+          <ApiKeyMetadataCard />
+          <ApiKeyEditableFieldCard
+            field="cidrs"
+            i18n={{ namespace: "firmaconio.createApiKey.form.network" }}
+          />
+          <ApiKeyEditableFieldCard
+            field="testers"
+            i18n={{ namespace: "firmaconio.createApiKey.form.network" }}
+          />
+        </CreateApiKeyForm>
       </Suspense>
     </Stack>
   );
